pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend="setuptools.build_meta"

[project]
name = "screenshot-to-text"
dynamic = ["version"]
requires-python = ">=3.11"
description= "Cross platform screenshot to text CLI"
readme = "README.md"
authors = [{name="Alphan Aksoyoglu", email="alphanbe@gmail.com"}]
dependencies = [
    "opencv-python-headless>=4.13.0.92",
    "platformdirs>=4.5.1",
    "pydantic>=2.12.5",
    "pypdf>=6.7.2",
    "pyyaml>=6.0.3",
    "tomli-w>=1.2.0",
    "typer>=0.22.0",
]

license = "MIT"
license-files = ["LICEN[CS]E*"]

[project.scripts]
screenshot-to-text="screenshot_to_text.cli:app"
s2t="screenshot_to_text.cli:app"

[tool.setuptools.packages.find]
where=["src"]
[tool.setuptools.package-data]
screenshot_to_text = ["data/*.yaml"]

[tool.setuptools.dynamic]
version = {attr = "screenshot_to_text.__version__.__version__"}

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.0",
]

[tool.ruff]
line-length = 160
